Extract share notification from notification API
