+++ backend/dataall/modules/datasets/services/dataset_location.py
@@ -1,5 +1,4 @@
 import logging
-from typing import List
 
 from sqlalchemy import and_, or_
