+++ backend/dataall/modules/datasets/tasks/subscription_service.py
@@ -13,6 +13,7 @@ from dataall.aws.handlers.sqs import SqsQueue
 from dataall.db import get_engine
 from dataall.db import models
 from dataall.modules.datasets.services.dataset_profiling_service import DatasetProfilingService
+from dataall.modules.datasets.services.share_notification_service import ShareNotificationService
 from dataall.tasks.subscriptions import poll_queues
 from dataall.utils import json_utils
 from dataall.modules.datasets.services.dataset_table import DatasetTableService
@@ -261,7 +262,7 @@ class SubscriptionService:
 
                             log.info(f'SNS update publish response {response}')
 
-                            notifications = db.api.Notification.notify_new_data_available_from_owners(
+                            notifications = ShareNotificationService.notify_new_data_available_from_owners(
                                 session=session,
                                 dataset=dataset,
                                 share=share_object,
